More closely align address translation with the specification.
The specification has a numbered list of steps in the algorithm.
The Sail version had implemented access permission checks (Steps 6,8)
before the superpage alignment check (Step 5).  This commit fixes
the ordering.

There is currently no functional difference between the two orders:
all the steps involved raise page-faults corresponding to the
access type; the only difference would be the error printed in
the execution log.

However, when shadow stack is implemented, it inserts a Step 7
that could raise an access-fault exception that would make the
difference in step order visible.

While here, mark in comments the steps of the algorithm for easier
correlation with the prose spec.