Revert "Fix minstreth behaviour"
This reverts commit cd54a3780178d8cf3ca0fa29009fa7f14059e014.

It hasn't been clarified in the ISA manual, but it has been clarified in [this Github issue](1) and in a new test in `riscv-tests` that the original behaviour was expected.

Essentially there are two different concepts: architectural CSRs (the instruction counter) and CSR addresses `minstret`, `minstreth`. The ISA manual conflates these concepts. When it says

> an explicit CSR write suppresses and overrides any implicit writes or modifications to the same CSR by the same instruction.

When it says CSR here it means the architectural CSR, not the CSR address.

[1]: https://github.com/riscv/riscv-isa-manual/issues/1255