Fix PMP address matching
In a recent refactor I accidentally removed the * 4 factor in `pmpaddr` registers. This fixes that.

I added a basic test that tests this at least for the NAPOT case. We could do with a more comprehensive test but it's quite difficult to write this sort of test as an ELF. I think maybe this is a situation where unit tests would be appropriate.

Finally I added a missing `sfence.vma` in the `test_max_pmp.c` test (doesn't actually matter for the Sail model but we should be correct), and simplified `crt0.S`'s full match PMP address.