Rename the variants used for instruction retirement.
The RETIRE_FAIL and RETIRE_{OK,SUCCESS} variants used in the return
type of the execute clause are increasingly confusingly named,
since now part of the execution and retirement now occurs in
the `pre_step()` state machine.

This renames them to `Exec_Incomplete` and `Exec_Complete`
respectively; this corresponds to their type already being
defined as `ExecutionResult`.  `Exec_Success` is defined as
to replace `RETIRE_SUCCESS` as a slightly shorter version of
`Exec_Complete()`.