Update Rocq termination information and add CI
The CI is based on a standard Rocq Docker image and uses the Sail release binary like the main CI job. It uses a separate workflow because it can be easily disabled if it causes trouble.

Providing an arbitrary limit for the non-terminating top-level loop is only useful for Rocq, not the Lean prover, so it is ifdef'd out.