Merge Retire_Failure into ExecutionResult
There's really no reason to have two levels of unions here so I added `Retire_Success` into `Retire_Failure` and then used that directly as `ExecutionResult`. This is definitely simpler and it also removes the weird case of WFI which was included in `Retire_Failure` when it isn't really failing to retire (especially if `wfi_is_nop()` is true).

I left the `RETIRE_SUCCESS` to make the commit smaller, but it does mean there is some stylistic inconsistency. We can fix that with a simple search and replace in future.