Add a project version to the build.
Uses `git describe` to get the version according to git,
and sanitizes it for CMake project function for potential
future use with CPack.

Adds an option to the c_emulator to report the (git) version.

This requires the github checkout action to fetch a clean
history at least upto the last tag.  For now, fetch all history.

For builds of a downloaded archive without git history, use
a specified version.