Add Decompress result for compressed instructions
The rationale for this is something we encountered when doing CHERIoT
verification using the generated SystemVerilog from Sail. By having the
compressed instructions directly invoke execute for the decompressed
variant, we end up with (for example) all the logic for load and store
instructions duplicated in the output.

It's also a bit awkward because it makes the main `execute` function
recursive, which inhibits some useful Sail optimisations.

This resolves this by adding a return type to the execute function
(enabled by some other recent commits).
```
Decompress : ast
```
which allows the top-level step function to handle executing the decompressed
variant.

As part of this PR that type is moved to where the `ast` type is declared
rather than being in a separate file.