Implement the Sdext debug mode extension.
The try_step API was changed to reflect the current debug mode
state and to provide a mechanism for the external harness to
request a change in that state (i.e. enter and exit debug mode).

Two new callbacks indicate entry and exit from debug mode.

This does not provide a model-external harness to connect to
external debugger infrastructure; that is left for the future.

Hook functions have been defined for behavior that is
UNSPECIFIED in the debug specification or depends on the
debugger's memory map (in riscv_sdext_hook_insts).

StopCount is implemented for the single-hart case, with a note
marking the change needed for multi-hart emulation.

This moves should_inc_minstret out of an extension-specific file
into riscv_platform in the core model, which fixes #783.

Some notes have been added under doc/ about the sdext
implementation.