Handle address translation for misaligned loads and stores better (#861)
Refactor the LOAD and STORE instruction so they split misaligned
accesses into multiple sub-accesses and perform address translation
separately. This means we should handle the case where a misaligned
access straddles a page boundary in a sensible way, even if we don't
yet cover the full range of possibilities allowed for any RISC-V
implementation.

There are options for the order in which misaligned happen, i.e. from
high-to-low or from low-to-high as well as the granularity of the
splitting,
either all the way to bytes or to the largest aligned size. The
splitting
can also be disabled if an implementation supports misaligned accesses
in hardware.

In addition tidy up the implementation in a few ways:

- Very long lines on the LOAD encdec were fixed by adding a helper

- Add some linebreaks in the code so it reads as less claustrophobic

- Ensure we use the same names for arguments in encdec/execute/assembly.
Previously we used 'size' and 'width'. I opted for 'width' consistently.

Co-authored-by: Alasdair Armstrong <alasdair.armstrong@cl.cam.ac.uk>

---------

Co-authored-by: Alasdair Armstrong <alasdair.armstrong@cl.cam.ac.uk>