Fix build failure due to plat_enable_htif signature mismatch (#930)
We compile the model as C, and unlike C++, in C prior to C23 a function
declaration with no parameters didn't mean that it was a function with
no parameters; instead it meant it was a function with unknown
parameters, so any would be allowed.

This was fixed in C23 and GCC 15 increased the minimum standard to C23,
thus exposing this mistake.