Fix macOS compilation and add to CI (#867)
Unfortunately since a Sail compiler binary for Mac is not yet available
this uses OPAM. Once one is available we'll switch to that.

This also fixes a dependency ordering error on all platforms -
`sail_runtime` depends on GMP but this wasn't declared so it could fail
if you didn't have GMP already available (i.e. you were building it from
source).

---------

Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>