Make RVFI a runtime option (#932)
Instead of compiling separate models/emulators for RVFI, compile a
single one with a run-time flag to enable RVFI.

The RVFI implementation is still in Sail, but it is now a runtime option
so an additional simulator doesn't need to be built.

Tested with TestRIG comparing to the previous, compile time RVFI version
of the model.