Use existing bit counting functions
Use the existing bit counting functions, and add `count_ones()`.

One downside is that `count_leading_zeros()` is not implemented in Sail, so it makes the inclusion in the ISA manual less helpful.

However given the difficulties of actually getting Sail code into the ISA manual it's probably ok for now. Also `count_leading_zeros()` is a standard operation and has obvious semantics.