Add Lean back to CI for each PR but don't require it to pass
Now that we have a "ci_pass" job, we can reenable the Lean CI job for each PR but not add it as a dependency of that job. This way we will know if Lean is broken by a new change and can hopefully get it fixed without bogging down developers who aren't familiar with the theorem prover backends.