Fix uses of to_bits_unsafe for bit rotate and Zvbb (#964)
This delegates the bits versions to the integer versions. This also
identified a dangerous looseness in the types - previously you could use
the bit version to shift by any amount, and it would give the wrong
results if the shift was more than `'n`.

In Zvbb we have to use `2 ^ SEW_pow` instead of `get_sew()` so that Sail
knows the link between `SEW` and `SEW_pow`.