Lean CI: use the latest version of sail (#987)
Before this change, we skipped sail build & install when hitting the
`.opam` cache. As a result, some random partially outdated sail version
was used, instead of the intended latest tip-of-tree of sail. This PR
continues to cache opam and sail, but does not skip the sail build &
install on a cache hit, but executes them, trusting that the cache will
accelerate them. As a result, we get the best of both worlds, fast
builds and always the latest sail version.