Support generating assembly for reserved vtypes
The `vset[i]vli` instructions take an immediate that sets the `vtype` CSR. The normal way to generate the immediate value is by giving values for all of `vtype`'s fields (`ma`, `ta`, `sew` and `lmul`).

`sew` and `lmul` have reserved values, however that does *not* mean that instructions that use those reserved values are necessarily reserved. Implementations are allowed to legalise the values (see the note in `handle_illegal_vtype()`).

Therefore we should support disassembling these instructions. The syntax for a `vset[i]vli` with a reserved `vtype` is just to use an integer immediate instead of the mnemonics.