Handle reserved nreg values for `vmv` in encdec
The `vmvNr.v` instruction only supports N of 1, 2, 4, or 8 (encoded as 0b00000, 0b00001, 0b00011, and 0b00111 respectively). Other values are reserved.

This changes the code to raise an illegal instruction exception directly in `encdec` instead of in the `execute` clause. Without this change it's possible to decode an opcode to `VMVRTYPE` but then not be able to disassemble it, because the `assembly` clause does not support the reserved values.

This also makes the code a bit simpler since it doesn't need to do bit twiddling tricks to get `N` - it comes directly from the decoder mapping.