Add a configuration validation function to the model. (#1005)
This is a start and can be extended to check more invariants of a valid
configuration.

Configuration validation is done when the model is initialized in
`init_model()`. `init_model()` and `reset()` are moved out of
`riscv_step.sail` into a new file `riscv_model.sail` to manage the
dependency on `riscv_validate_config.sail`.

Addresses #860.