Improve type checking performance for generate_isa_string
Currently Sail is can be quite slow when resolving the type for expressions
containing many overloaded operators, particularly when the existing heuristics
for speeding this process up do not fire. `The generate_isa_string` function
hit this issue, increasing the time it takes to check the model significantly.

This PR resolves this by introducing an `append_strings` helper that allows
writing `"a" ^ "b" ^ "c"` as `append_strings(["a", "b", "c"])`.