Use `0` for empty pred/succ set in fence assembly (#1030)
It's legal (but not very useful) to have a fence with an empty
predecessor or successor set. These are executed as nops. Currently we
disassemble them to e.g. `fence , rw` which is not valid. This improves
the disassembly to `fence 0, rw` which is accepted by LLVM, but
unfortunately not GAS.