riscv_device_tree: refactor generate_isa_string
The sail compiler shouldn't blow up when processing the extension
vector, and splitting by categories is not that important when we merge
them back shortly after, so use comments to categorize instead.

This makes the code a bit more readable to me.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>

riscv_extensions: move the ordered vector of extensions

Having the ordered extension vector closer to the definition of
extensions should make it harder to forget about the isa-string.

I would still like to generate the correctly ordered list from the enum.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>