Add missing SEW_widen checks for Zvbb (#1057)
These instructions were missing checks that the EEW (`SEW_widen`) is not
greater than 64. The spec states:

> The V extension supports EEW of 8, 16, and 32, and 64.

And also

> Note: If Zve32x is supported then Zvkb or Zvbb provide support for EEW
of 8, 16, and 32. If Zve64x is supported then Zvkb or Zvbb also add
support for EEW 64.

So eventually we will need to make this configurable, but it never
supports >64 bits anyway.