Improve typing of division and modulo
This changes the definition of `%` and `/` to *only* work with positive numbers and also require the divisor to be non-zero.

The very few number of places that need signed division can call `quot_round_zero`/`rem_round_zero` directly. These don't require positive arguments, but they do still require a non-zero divisor.

Making this change required updating a load of other code which didn't have sufficiently accurate types, so it's a quite a large commit but not easy to split up.