Remove speculate_conditional (#1060)
I've previously used this to inject sprurious Store-Conditional
failures, however it required some tweaking because it's called very
early before various exceptions can happen.

A better solution is to spuriously cancel the reservation directly in
`load_reservation()` or by calling `cancel_reservation()`.

This change doesn't actually do that but I added notes explaining that
you can. In future when we provide a library interface we can do it like
that.