Fix a few issues in the `vset{i}vl{i}` instructions. (#1051)
- Check for reserved values of `vsew` and `vlmul` in the proposed
`vtype` before setting `vtype`. If reserved values are present, set
`vtype.vill` instead of asserting (as was the case for all three
instructions).

- Check that the entire proposed `vtype` is valid, and ensure that the
reserved bits of `vtype` are zeroed as required instead of blindly
copying them from the proposed `vtype` (as was the case for `vsetvl`).

-  Check valid vtype condition when rs1=rd=x0.

- Ensure `vstart` is zeroed when returning due to an illegal proposed
`vtype`. Add a note as a reminder that this only for the non-trapping
case.

Fixes #1095.