Add config flags to control PMP address mode support (#1114)
This adds config flags to allow marking PMP address matching modes as
not supported. If software tries to write an unsupported mode it will
get set to OFF instead.

For NA4 mode it is required that the grain G is 0. If it isn't then that
mode is unsupported. This is also checked in `config_is_valid()` which
is redundant... but I think that's ok for now. Another option would be
to print a warning like `Warning: NA4 is marked as supported but G > 0
so it has been disabled`.

Fixes #1109