Add Remote Bitbang protocol support for GDB/OpenOCD integration
This PR adds the Remote Bitbang protocol (mainly code from Spike), which is one of three components required to enable remote debugging (Remote Bitbang, the Debug Module and the Sdext Extension).

With this protocol in place, it's now possible to connect OpenOCD (compiled with Remote Bitbang support) to the simulator. OpenOCD will successfully initiate communication but will not complete the initialization process yet, as the Debug Module is still missing. That module will eventually provide registers and support for accessing general-purpose registers and CSRs, both of which are not yet implemented.

To enable remote debugging, this PR also adds a new (internal for now) command-line option to Sail:

./build/c_emulator/sail_riscv_sim -d 9824 ./test/riscv-tests/rv64ui-p-jal.elf

The -d <PORT> option sets the port for OpenOCD to connect to. Sail will wait for an incoming connection, but will continue the simulation slowly.