Add support for Sail unit tests (#710)
This adds basic support for Sail unit tests. These have some pros and
cons over C tests:

Pros:

* They don't require a RISC-V C compiler.
* They can test internal code (hence why I've called them unit tests).
* They can "cheat", e.g. to change privilege mode you can just magically
`cur_privilege = Supervisor`. With C you have to implement a syscall,
etc.

Cons:

* They are tied to internal code which will make refactoring more
tedious because it means you are likely to need to update the tests. We
can minimise that by using "high level" functions like `execute()`,
`read_CSR()`, etc.
* They only run on the Sail model, so we can't e.g. verify them against
SPIKE.

Given the sorry state of testing in this repo, and the fact that we
don't have a way of writing C tests at all yet, I think this is probably
a good medium term approach.

Currently the unit test executable reuses a load of code from the
emulator, which is not ideal. When we have wrapped the model in a nice
C++ library we can use that instead.