Move privilege-based CSR access into scattered is_CSR_accessible function
Instead of having special extra functions to check accessibility of CSRs that depend on the privilege mode, and `is_write`, just add those parameters to `is_CSR_defined` (and rename it to `is_CSR_accessible`).

This also fixes a bug (discovered by inspection when making this change) - the `check_Counteren` was not checking the `...h` registers for RV32.