Replace `getopt` CLI processing with CLI11.
The CLI help and error processing has been improved; in particular,
the help now shows the tracing options.

CLI11 (v2.5.0) is added as a dependency, and pre-commit has been
tuned to not format files under dependencies.  The CLI11 LICENSE
is added to the installed files.

Light customization of the help message formatting has been done
to get the trace options in the footer and to remove redundant
information about positional options.

Backwards incompatible changes:
- the ABI name option is moved from `-[vV] abi` to a top-level
  option `-A`, since `abi` was not really a kind of tracer but
  a tunable for the `reg` tracer
- `-[vV]` require an explicit argument; this breaks previous uses
  of these options that took the absence of an argument to imply
  `all`