Add sail_project file to modularise the specification (#572)
Use the Sail module system to organise the files in the model. Rather
than passing a list of files on the command line, the files are
specified using the `model/riscv.sail_project` file.

The previous version uses a big flat list of files which is unwieldy and
hard to understand as it is not clear immediately what the high level
structure of the model is, and the way we configure the model does not
allow for extensions to easily be added and removed without
hand-modifying the CMake. This doesn't matter for the C backend but it
does for formal/SV backends.

To solve this problem, I have added a basic module system to Sail, which
allows for the structure of the specification to be specified as a
collection of modules in a `.sail_project` file. The key ideas here are
as follows:

- Each of these modules consists of a list of files (much like how we
passed a list of files to Sail on the command line previously), as well
as a list of modules that module requires.

- The module system ensures that if a module does not require another
module, it cannot refer to any definitions contained within it. There is
also a `private` keyword which can be used to encapsulate implementation
details.

- Modules can specify other modules that must come before or after them,
but without requiring them. This is to support the scattered definition
feature.

- The `.sail_project` file serves as an instant reference to the overall
model structure, it isn't embedded in the files. You can also have
multiple `.sail_project` files, so you could pass `riscv.sail_project
cheri.sail_project` to Sail, and it would be able to link them together
in much the same way we do now.

- There is some basic support for conditional includes.

With this, building the model becomes as simple as:

sail --project model/riscv.sail_project --variable ARCH=RV64
--all-modules

---------

Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>