Fix softfloat debug builds (#1176)
I'm not sure exactly why softfloat is set up like this, but with
`INLINE` only defined as `inline`, then it was possible in debug builds
to get undefined references to some of the inlined functions (e.g.
`softfloat_countLeadingZeros32`). Apparently in C99 `inline` functions
*might* not actually get emitted; you need to use `static inline`. I
don't fully understand why but ChatGPT called this "The C99 Inline Mess"
when I asked about it so probably not worth digging!

The old version of `platform.h` also used `static inline`.