Improve command-line argument parsing and help formatting (#1129)
Replace `getopt` CLI processing with CLI11.

The CLI help and error processing has been improved; in particular, the
help now shows the tracing options.

CLI11 (v2.5.0) is added as a dependency, and the CLI11 LICENSE is added
to the installed files.
   
Backwards incompatible changes:
- short options have been removed
- the ABI register name option has moved from `-[vV] abi` to a top-level
option `--use-abi-names`, since `abi` was not really a kind of tracer
but a tunable for the `reg` tracer
- tracing is now off by default and there are separate `--trace-mem`,
`--trace-instr`, etc. flags as well as `--trace-all`.

---------

Co-authored-by: Prashanth Mundkur <103736+pmundkur@users.noreply.github.com>
Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>