Improve vector nfields handling (#1165)
Instead of storing the raw bits in the decoded `instruction` store the
actual integer value of `nfields`. This also allows us to handle the
fact that non-power of two integers are reserved for vector whole
register load/stores (`VLRETYPE`/`VSRETYPE`). Previously that was
handled in the `execute()` clause, which is unfortunate because then we
can decode those reserved instructions but not disassemble them. (And we
probably shouldn't be decoding instructions that are *always* reserved
anyway.)