Decouple callbacks impl from model by introducing callback interface (#1146)
Currently the callbacks implementation depends directly on the RVFI
implementation. This commit avoids that by introducing a callback
registration mechanism, allowing multiple targets to receive the
callbacks. Currently the targets are text logs, and RVFI. In future we
will add a binary log format, and probably a C lockstep verification
API.

---------

Co-authored-by: Prashanth Mundkur <prashanth.mundkur@gmail.com>
Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>