Fix `read_vreg` for negative LMUL in widening vector instructions. (#1179)
Call the single register read only when all the (potentially wider)
elements fit into `vlen` bits.

This fixes a bug with widening instructions with negative LMUL, where a
negative LMUL was assumed to imply that a single register read sufficed.
In this situation, using `num_elems` in conjunction with the widened SEW
could cause out-of-bounds read-accesses into the register's bits.

---------

Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>