Implement JSON schema validation for the configuration.
This uses an externally installed [boon](https://github.com/santhosh-tekuri/boon)

For now, it needs to be explicitly enabled with `--validate-config-schema`.
With this flag, _only_ schema conformance is checked; that is, `--validate-config`
is still needed to validate other aspects of the configuration.

The schema is embedded into the executable, removing the need
to locate the schema file at run time.