Implement JSON schema validation for the configuration.
This uses an externally installed [boon](https://github.com/santhosh-tekuri/boon)

For now, it needs to be explicitly enabled with `--validate-config-schema`.
With this flag, _only_ schema conformance is checked; that is, `--validate-config`
is still needed to validate other aspects of the configuration.

If not specified, the schema file is looked for in a few places (the build
directory or its standard installation location), otherwise the
`--config-schema` flag needs to be used to specify its location.