Don't use generated reset vector to initialise registers (#1220)
Instead of loading and executing actual instructions into the model to
set a0=mhartid, a1=&dtb, we can just do that directly in the model. The
DTB is always loaded at the start of the ROM so we don't even need to
separately tell the model where it is.

This has the advantage that no "extra" instructions are executed before
the ELF's which is less confusing and simpifies testing against other
simulators.

Finally I simplified how the DTB is actually loaded, and improved the
printing of exceptions.