Return softfloat flags and result as a tuple (#1275)
Since we now use a proper header for the generated model code, we can
#include it in `riscv_softfloat.h` and use the `(fflags, bits(..))`
types to directly return the tuples that the rest of the Sail code wants
instead of using the super hacky method of poking some registers inside
the model.

This is much much cleaner, makes the softfloat functions pure, and
removes their dependency on the model *implementation*. Now it only
depends on the model types (and that is really only because C doesn't
have native tuple types; with fully native C++ codegen we could
theoretically have them return `std::tuple<...>`).

I also switched `riscv_softfloat.c` to C++ which has slightly nicer
syntax for returning values (and is more consistent with everything else
being C++).