Improve float classify code
Extract float classification into a function and use an enum for the result. This reduces code duplication and the presence of magic numbers.

This is the first use of Sail's built-in float library, which we should try to migrate to.

I'm on the fence about `num_of...` being a bit too magical. Maybe we should just use a mapping to `bits(10)`.