Fix RMEM file selection (#1281)
The RMEM backend uses a different definition of the `jalr` execute
clause and expects extra instructions to be defined for internal use.
That distinction was accidentally dropped when we switched to the Sail
project file. This introduces a variable in the project file to handle
the different file lists the same way we handle the termination file for
Lean/Rocq.

The RMEM backend is still broken for other reasons, but the other error
messages are the same with or without this change.