Improve float classify code (#1288)
Use the Sail library's new float classify function. This reduces code
duplication and the presence of magic numbers.

This is the first use of Sail's built-in float library, which we should
try to migrate to.

The SMT properties should ideally live in the Sail repo but it doesn't
have SMT testing set up yet.