Add Zic64b extension (#1351)
This extension does not have a functional impact on the model and in
practice just limits the legal value of the `cache_block_size_exp`
config option. It is listed as one of the extensions in the RVA23
profile and gcc accepts it as part of `-march` though, so I think it is
worth including in the model. It is helpful validation for someone
intending to create an RVA23 configuration.