Add more state change callbacks (#1354)
This adds a few state change callbacks that we needed for an internal
binary logging format. They aren't used by the RVFI or log callback
handlers but they should also be useful for the open source binary
logging format that we eventually add (#545).

* Add a `fetch_callback` that says which instruction was executed (if
any).
* Add `redirect_callback` that is called for taken and unconditional
branches.
* Add `is_interrupt` and `cause` to `trap_callback`.

I also renamed `value` to `new_pc` to be a bit more explicit.

Fixes #1352