Add an option to choose a smaller set of modules for Lean build (#1219)
Add a `SAIL_MODULES` variable to the CMakeLists file that allows
specifying a smaller set of modules, useful to constraint the Lean build
to be something like RV64IM or similar, when that's all we need for some
use cases. For example:
```
 cmake -S . -B build -DCMAKE_BUILD_TYPE=RelWithDebInfo -DSAIL_MODULES="I;M;riscv_termination;riscv_main"
```
Not sure if this is the best way to do this with CMake, or if we want to
make it work more broadly other than just for Lean.