Allow configuring read-only-zero PMP entries. (#1363)
According to the spec: `All PMP CSR fields are WARL and may be read-only
zero.`

This adds a `memory.pmp.usable_count` parameter that specifies which of
the `memory.pmp.count` PMP entries are usable. Entries lower than the
`usable_count` are writable, higher entries are read-only zero.

This handles probably the most common configuration of read-only zero
PMP entries, and fixes #1111.