Add Sstvecd to supported extensions. (#1377)
This just captures already supported functionality in `stvec`. For now,
this extension is just gated on support for the Supervisor mode. When we
add granular configuration options for modes in `stvec`, this extension
can be exposed as a configuration option.

Sstvecd is a profile-defined extension coming from RVA20S64. It's not
clear what version this extension is, so I'm just using v1.0.