Add ExecuteAs result for compressed instructions. (#1382)
The rationale for this is something we encountered when doing CHERIoT
verification using the generated SystemVerilog from Sail. By having the
compressed instructions directly invoke execute for the decompressed
variant, we end up with (for example) all the logic for load and store
instructions duplicated in the output.

It's also a bit awkward because it makes the main `execute` function
recursive, which inhibits some useful Sail optimisations.

This resolves this by adding a return type to the `execute` function
```
ExecuteAs : instruction
```
which allows the top-level step function to handle executing the
decompressed variant.

Removing the recursion in `execute` also allows the removal of the
associated termination measures.

For the Lean build, `BEq` derivations have been suppressed for
`ExecutionResult` and `Step` since they now embed an `instruction`, and
the derivation for the latter was suppressed for performance reasons.

---------

Co-authored-by: Alasdair Armstrong <alasdair.armstrong@cl.cam.ac.uk>