Add option not to download jsoncons (#1385)
Replaces a previous hack to avoid a download in the Rocq opam package
for compatibility with recent versions of cmake.

An alternative to this would be to use `ExternalProject` in the same way
that the GMP support works, but that would be a little more involved and
change the timing of the download from configure-time to build-time.