Greatly simplify the code for reading & writing vector registers/masks (#1213)
The original code was very over-complicated.

Where `assert(-3 <= EMUL_pow & EMUL_pow <= 3);` is added, the
calculation of EMUL_pow is clearly wrong - see #1196. Since this was an
existing issue I have just added a TODO rather than fixing it.

This also introduces `static_assert()` which can be used for documenting
properties that are already known by the type checker to be true (and it
will give nice error messages if you edit the code so that it is no
longer true).