Add support for Sstvala.
This requires differentiating between hardware and software breakpoints.
An enum was added for this; the enum for hardware breakpoints would be
used for breakpoint exceptions caused by triggers and the debug module.

The non-zero xtval settings have been hoisted into `core/platform_config`.

Although Sstvala strictly applies to only `stval`, and there are also
Shtvala and Shvstvala for `htval` and `vstval`, it is simpler to group
them under `xtval` for now.