Slight refactor of address translation and PTE checks to ease implementing Zicfiss. (#1393)
- This explicitly separates Steps 6 and 8 of the address translation
algorithm, so that the Zicfiss support in Step 7 can be cleanly
inserted.

- The PTE checking API was adjusted. Failures in steps 6 and 8 both
result in page fault exceptions, and this failure cause was implicitly
assumed. However, failure in Step 7/Zicfiss results in an access fault,
necessitating a need to be more explicit. Adjusting for this
explicitness has a minor impact on the `ext_` API used for the CHERI PTE
and PTW hooks. The documentation for these hooks has been updated.

Comments have been left to indicate the changes for Zicfiss.