Use C++ Sail output (#1274)
This switches from using Sail's C output to using C++. The output code
gets wrapped in a `class`, which means we can create more than one
instance of it (e.g. for multicore).

The `Model` class is currently *not* thread safe due to the use of
temporary globals in `sail.c`. So although you can simulate multicore
systems, you have to execute them one at a time.

The handling of platform callbacks is a little convoluted but I think
this is a quite flexible solution. Sail is instructed to derive the
`Model` class from `PlatformImpl` which has virtual methods for all of
the platform callbacks with default nop implementations.

This means the default `Model` can be constructed and has a "nop"
platform. Then we add our implementation of the platform with
`ModelImpl` which overrides those methods.

`ModelImpl` also allows registering callback receivers.

This is not 100% perfect yet. We still have a global `g_model`, which is
not ideal.

---------

Signed-off-by: Tim Hutt <tdhutt@gmail.com>
Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>
Co-authored-by: Nadime Barhoumi <nadime@riscv.org>