Replace hardcoded constant in tlb_hash() with a compile-time constant
Replaced hardcoded constant in tlb_hash() with tlb_entries_bits to improve maintainability.
This ensures that if the TLB size is changed in the future, the hash computation will
automatically adapt to the new number of entries.

Preferably I would have not introduced new type and used log2(), but Sail does not support it.