Fix location of the `hartSupports_measure` function.
It was defined in a location that seems to have preceded the
the compiler's knowledge of all the variants of the `extension`
enum, resulting in unused variable warnings from the latest Sail
compiler:
```
Warning: Redundant case core/extensions.sail:25.4-15:
25 |    Ext_Sstvecd => 1, // > S
   |    ^---------^
   |
This match case is never used

Warning: Redundant case core/extensions.sail:26.4-15:
26 |    Ext_Ssu64xl => 1, // > S
   |    ^---------^
   |
This match case is never used
```
and so on, due to the first match pattern `Ext_D` being treated
as a variable that made all subsequent matches redundant.

This moves the function next to `currentlyEnabled_measure` and adds
comments.