Improved clang-format style (#1396)
This fixes various annoyances in our clang-format style that have been
bothering me for a while:

* Binpacking arguments/parameters and aligning them to the opening `(`
is *really* bad for diffs/conflict resolution. This changes it to try to
match the "prettier" algorithm (also used by Sail's autoformatter),
which has much simpler and better behaviour - if the arguments fit on
one line do that, otherwise put each argument on its own line.
* Don't put functions on one line. This is also a little diff-unfriendly
and feels unnecessarily inconsistent to me.
* The line length limit of 80 characters is very constrictive so I
increased it to 120 characters.
* For some reason it didn't put `{` on a new line... except for function
bodies, which is weirdly inconsistent. Now `{` never starts on a new
line.
* Sort `#include`s. This might have been needed in the past but I
verified it builds with sorted includes now.