Add JSON output format
This adds support for outputting an execution trace using [JSON-Lines](https://jsonlines.org/), i.e. newline-separated JSON objects.

This isn't an ideal format, but it is at least universally supported and much better than parsing the text-based output which was never intended to be machine readable.

In future I hope we will add a few other formats - RVVI-TEXT, Protobuf, etc. but I think JSON is a bare minimum.

The schema is ad-hoc but quite simple.