Add reserved_behavior.amocas_odd_registers config (#1403)
Add a configuration option to control the behaviour when executing
AMOCAS instructions that are reserved due to the use of odd register
indices.

This is the first setting we've added to control reserved behaviour. It
allows two options: a fatal error that stops the simulation, or to treat
it as an illegal instruction (the default choice).

See  #775

---------

Signed-off-by: Tim Hutt <tdhutt@gmail.com>
Co-authored-by: Tim Hutt <tdhutt@gmail.com>
Co-authored-by: Prashanth Mundkur <prashanth@riscv.org>